+++ scripts/dev/activate_venv.ps1
@@ -7,6 +7,7 @@ $activateScript = "$venvPath\Scripts\Activate.ps1"
 if (Test-Path $activateScript) {
     # .venv exists, activate it and show "next steps" message
     & $activateScript
+    Clear-Host
     Write-Host ""
     Get-Content -Path (Join-Path $PSScriptRoot "next_steps.txt") | Write-Host
     Write-Host ""